Move the widget to a separate file in "includes" folder
